requirements.txt
python-telegram-bot==21.10
anthropic
aiohttp
pytz
SpeechRecognition
pydub
gTTS
